--- Project-2.Models/Favorite copy.cs
@@ -1,23 +1,0 @@
-using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
-
-namespace Project_2.Models;
-
-[Table("Offers")]
-public class Offer(Guid UserID, Guid PropertyID, string BidAmount)
-{
-    [Key]
-    public Guid FavoriteID { get; set; } = Guid.NewGuid();
-    [Required]
-    [ForeignKey("UserID")]
-    public Guid UserID { get; set; } = UserID;    
-    [Required]
-    [ForeignKey("PropertyID")]
-    public Guid PropertyID { get; set; } = PropertyID;
-
-    [Required]
-    [StringLength(50)]
-    public string BidAmount { get; set; } = BidAmount;
-
-    public DateTime Date { get; set; } = DateTime.Now;
-}
